rename SearchColumns.ts
